+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added `CheckNullifiersByPrefix` endpoint (#419).
 - Support multiple inflight transactions on the same account (#407).
 - Added `SyncNotes` endpoint (#424).
+- Cache sql statements (#427).
 
 ### Fixes